Add login and logout buttons to top bar

+++ app/src/main/java/com/example/mytradingapp/model/TradeItem.kt
@@ -5,13 +5,13 @@ import java.io.Serializable
 data class TradeItem(
     val id: Int = 0,
     val description: String,
-    val price: Double,
+    val price: Int,
     val sellerEmail: String,
     val sellerPhone: String,
     val time: Int = 0,
     val pictureUrl: String = ""
 ) : Serializable {
-    constructor(description: String, price: Double, sellerEmail: String, sellerPhone: String) :
+    constructor(description: String, price: Int, sellerEmail: String, sellerPhone: String) :
             this(0, description, price, sellerEmail, sellerPhone)
 
     override fun toString(): String {
